Security Patch for edX: 2016-12-7 We just released a patch (which is attached to this email) for a security vulnerability with login and registration. The attached patch should apply to both release of edxapp and Eucalyptus.
While the forms in question were typically submitted as POST requests via JavaScript, in rare cases we have observed that some requests were submitted as GET requests. This was possible because certain forms did not explicitly specify a method attribute. As a result, plain text passwords were passed in the query string and stored in the logs. While the traffic was encrypted in transit, users' passwords would have been available to system administrators or anyone with access to HTTP server logs.